+++ AppWebMvc/AppWebMvc/Controllers/AdminBlogController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AppWebMvc.Controllers
+{
+    public class AdminBlogController : Controller
+    {
+        public IActionResult Index()
+        {
+            // Admin'in Blog Yönetim Action'u
+            return View();
+        }
+    }
+}
