--- AppWebMvc/AppWebMvc/Controllers/AdminControllers/AdminBlogController.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace AppWebMvc.Controllers.AdminControllers
-{
-	public class AdminBlogController : Controller
-	{
-		public IActionResult Index()
-		{
-			// Admin'in Blog Yönetim Action'u
-			return View();
-		}
-	}
-}
