--- AppWebMvc/AppWebMvc/Controllers/AdminControllers/AdminHomeController.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace AppWebMvc.Controllers.AdminControllers
-{
-	public class AdminHomeController : Controller
-	{
-		
-		public IActionResult Index()
-		{
-			// Admin'in AnaSayfa Action'u
-			return View();
-		}
-	}
-}
