--- AppWebMvc/AppWebMvc/Controllers/AdminControllers/AdminSettingController.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace AppWebMvc.Controllers.AdminControllers
-{
-	public class AdminSettingController : Controller
-	{
-		public IActionResult Index()
-		{
-			// Admin'in Ayarlar Action'u
-			return View();
-		}
-	}
-}
