+++ AppWebMvc/AppWebMvc/Controllers/AdminPageController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AppWebMvc.Controllers
+{
+    public class AdminPageController : Controller
+    {
+        public IActionResult Index()
+        {
+            // Admin'in Sayfa Yönetim Action'u
+            return View();
+        }
+    }
+}
