+++ AppWebMvc/AppWebMvc/Controllers/AdminSettingController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AppWebMvc.Controllers
+{
+    public class AdminSettingController : Controller
+    {
+        public IActionResult Index()
+        {
+            // Admin'in Ayarlar Action'u
+            return View();
+        }
+    }
+}
