+++ AppWebMvc/AppWebMvc/Controllers/UserBlogController.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AppWebMvc.Controllers
+{
+    public class UserBlogController : Controller
+    {
+        public IActionResult Search(string? query, int? page)
+        {
+            // User'ın Blog Arama Action'u
+
+            return View();
+        }
+
+        public IActionResult Detail(int? id)
+        {
+            // User'ın Blog Detay Action'u
+            return View();
+        }
+    }
+}
