+++ AppWebMvc/AppWebMvc/Controllers/UserHomeController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AppWebMvc.Controllers
+{
+    public class UserHomeController : Controller
+    {
+        // User'ın AnaSayfa Action'u
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
